Quotient Filter Element edits

+++ quotient_filter/quotient_filter_element.cpp
@@ -1,6 +1,6 @@
 #include "quotient_filter_element.h"
 
-QuotientFilterElement::QuotientFilterElement(int value, bool is_occupied, bool is_continuation, bool is_shifted) {
+QuotientFilterElement::QuotientFilterElement(long long int value, bool is_occupied, bool is_continuation, bool is_shifted) {
     this->value = value;
     this->is_occupied = is_occupied;
     this->is_continuation = is_continuation;
